% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plots.R
\name{donut_plot}
\alias{donut_plot}
\title{Create a donut plot}
\usage{
donut_plot(filtered.vcf)
}
\arguments{
\item{variants.df}{is the filtered and re-formatted dataframe from filtered VCF from filter_vcf() function.}
}
\value{

}
\description{
Create a donut plot
}
\examples{
\dontrun{
vcf <- vcfR::read.vcfR("/path/to/vcf")
vcf.df <- cbind(as.data.frame(vcfR::getFIX(vcf)), vcfR::INFO2df(vcf))
vcf.s4 <- VariantAnnotation::readVcf("/path/to/vcf")
variants.df <- filter_vcf(vcf.df,vcf.s4)

donut_plot(filtered.vcf=variants.df)
}

}
