% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_txDB.R
\name{download_gen_refs}
\alias{download_gen_refs}
\title{Create TxDB and Genes GRanges object from a GFF file}
\usage{
download_gen_refs(
 
    URL = "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_22/gencode.v22.annotation.gff3.gz",
  save = FALSE
)
}
\arguments{
\item{URL}{The URL of the reference genes in gff3 format}

\item{save}{boolean - whether to save the txDB as sqlite and the}
}
\value{

}
\description{
Create TxDB and Genes GRanges object from a GFF file
}
\examples{
objects <- download_gen_refs()

head(objects$gtf)
objects$txdb

}
