% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_reformat_vcf_df.R
\name{filter_ctat_vcf}
\alias{filter_ctat_vcf}
\title{Filter and Reformat the VCF from v2.5 CTAT Mutations Pipeline}
\usage{
filter_ctat_vcf(vcf.df, vcf.s4)
}
\arguments{
\item{vcf.df}{is a dataframe derived from vcfR package}

\item{vcf.s4}{is a s4 vectors object from bioconductor VariantAnnotation package}
}
\value{

}
\description{
Filter and Reformat the VCF from v2.5 CTAT Mutations Pipeline
}
\examples{
\dontrun{
vcf <- vcfR::read.vcfR("/path/to/vcf")
vcf.df <- cbind(as.data.frame(vcfR::getFIX(vcf)), vcfR::INFO2df(vcf))
vcf.s4 <- VariantAnnotation::readVcf("/path/to/vcf")
vcf.filtered <- filter_ctat_vcf(vcf.df,vcf.s4)
}

}
