% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz_tracks.R
\name{gene_tracks}
\alias{gene_tracks}
\title{Title}
\usage{
gene_tracks(vcf_s4, txdb, genome = "hg38")
}
\arguments{
\item{vcf_s4}{is a s4 vectors object from bioconductor VariantAnnotation package}

\item{txdb}{is a transcript database from}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
txdb <- GenomicFeatures::makeTxDbFromGFF("gencode.v22.chr_patch_hapl_scaff.annotation.gtf.gz")
AnnotationDbi::saveDb(txdb, "gencode.v22.chr_patch_hapl_scaff.annotation.sqlite")
vcf.s4 <- VariantAnnotation::readVcf("/path/to/vcf")
gene_tracks(vcf_s4 = vcf.s4[1:3], txdb=txdb)
}

}
