% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanMine_query_functions.R
\name{query.Disease}
\alias{query.Disease}
\title{Query HumanMine for associated diseases}
\usage{
query.Disease(gene.symbol)
}
\arguments{
\item{gene.symbol}{character(1) string, specifying gene symbol}
}
\value{
Dataframe of associated diseases
}
\description{
Query HumanMine for associated diseases
}
\examples{
query.Disease("SCNN1A")
}
