% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanMine_query_functions.R
\name{query.Expression}
\alias{query.Expression}
\title{Query HumanMine for tissue expression data}
\usage{
query.Expression(gene.symbol)
}
\arguments{
\item{gene.symbol}{character(1) string, specifying gene symbol}
}
\value{
Dataframe of tissue expression data
}
\description{
Query HumanMine for tissue expression data
}
\examples{
query.expression("SCNN1A")
}
