% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanMine_query_functions.R
\name{query.Publication}
\alias{query.Publication}
\title{Query HumanMine for publications relevant to genes of interest}
\usage{
query.Publication(gene.symbol)
}
\arguments{
\item{gene.symbol}{character(1) string, specifying gene symbol}
}
\value{
Dataframe of publications and relevant data
}
\description{
Query HumanMine for publications relevant to genes of interest
}
\examples{
query.Publications("SCNN1A")
}
