% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanMine_query_functions.R
\name{query.Snps}
\alias{query.Snps}
\title{Query HumanMine for snps in genes of interest}
\usage{
query.Snps(gene.symbol)
}
\arguments{
\item{gene.symbol}{character(1) string, specifying gene symbol}
}
\value{
Dataframe of snp data for genes of interest
}
\description{
Query HumanMine for snps in genes of interest
}
\examples{
query.Snps("SCNN1A")
}
