% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_get_and_filter_function.R
\name{query_get_and_filter}
\alias{query_get_and_filter}
\title{Query HumanMine using gene symbols}
\usage{
query_get_and_filter(gene.names)
}
\arguments{
\item{gene.names}{character(1) string, specifying gene symbols to query}
}
\value{
Disease, tissue, snps, and publications associated with genes of interest
}
\description{
Query HumanMine using gene symbols
}
\examples{
query_get_and_filter(c("SCNN1A", "SCNN1B", "SCNN1G", "SCNN1D"))
}
