% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set-up the project directories}
\usage{
setup(sample_name, directory = "reports")
}
\arguments{
\item{sample_name}{sample_name is a character string for the sample identifier, which must be included in the VCF filename}

\item{directory}{a character string for the name of the directory where reports will be saved.}
}
\value{

}
\description{
Set-up the project directories
}
\examples{
setup("my_reports")
}
